Is a Space-Filling Design Available?

Determine if a design from https://www.spacefillingdesigns.nl/ is available in this package based on how many characteristics (i.e. parameters), size (i.e., number of grid points), and type.

sfd_available(num_param, num_points, type = "any")

Arguments

num_param

An integer between two and ten for the number of characteristics/factors/parameters in the design.

num_points

An integer for the number of grid points requested. If there is no corresponding design, an error is given (when using get_design())

type

A character string with possible values> "any", "Audze-Eglais", "max_min_l1", and "max_min_l2", A value of "any" will choose the first design available (in alphabetical order).

Value

A logical

Examples

sfd_available(2, 10)
#> [1] TRUE
sfd_available(2, 10^5)
#> [1] FALSE